Generate test nodes from .kitchen.yml suites
Monkey patch CommonSandbox class to generate test nodes from
.kitchen.yml suites. Generated nodes can be easily overrided by users
by creating node files, as without this patch.

We use our previous patch of RunAction to declare a global variable
containing the suites configuration, so we are able to access it in
CommonSandbox.

Add some comments on our monkey classes (test/*.rb) and use require
instead of load to use them.